chore: pull out resolve_used_table_names and extract_manifest
