+++ wren-core-py/src/lib.rs
@@ -4,6 +4,7 @@ use remote_functions::PyRemoteFunction;
 
 pub mod context;
 mod errors;
+mod extractor;
 mod manifest;
 pub mod remote_functions;
 
@@ -14,5 +15,8 @@ fn wren_core_wrapper(m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<context::PySessionContext>()?;
     m.add_class::<PyRemoteFunction>()?;
     m.add_class::<manifest::PyManifest>()?;
+    m.add_function(wrap_pyfunction!(extractor::py_resolve_used_table_names, m)?)?;
+    m.add_function(wrap_pyfunction!(extractor::extract_manifest, m)?)?;
+    m.add_function(wrap_pyfunction!(manifest::to_json_base64, m)?)?;
     Ok(())
 }
